requirements.txt
biopython>=1.69
docopt>=0.6.2
networkx>=1.11
numpy>=1.12.1
pandas>=0.20.1
PubChemPy>=1.0.4
pypdb>=1.103
xmltodict>=0.11.0